Finish zlib decompression support

+++ weechat-protocol.c
@@ -42,7 +42,7 @@ static gchar* wtype_to_gvtype(const gchar* wtype, gboolean maybe)
 
 // TODO: Test me.
 static GVariant* weechat_decode_from_arg_to_gvariant(GDataInputStream* stream,
-                                                     type_t type, gboolean maybe, gsize* remaining)
+                                                     type_t type, gboolean maybe, gint32* remaining)
 {
     GVariant* val;
     gchar* s_type = g_strdup(types[type]);
@@ -152,14 +152,14 @@ error_free:
 void weechat_receive(weechat_t* weechat)
 {
     answer_t* answer = weechat_parse_header(weechat);
-    gsize remaining = answer->length - 5;
+    gint32 remaining = answer->length - 5;
     GDataInputStream* mem;
 
     if (answer->compression == 1) {
         mem = g_data_input_stream_new(
             g_converter_input_stream_new(
-            g_memory_input_stream_new_from_data(answer->body, remaining, NULL),
-            g_zlib_decompressor_new(G_ZLIB_COMPRESSOR_FORMAT_ZLIB)));
+                g_memory_input_stream_new_from_data(answer->body, remaining, NULL),
+                g_zlib_decompressor_new(G_ZLIB_COMPRESSOR_FORMAT_ZLIB)));
     } else {
         mem = g_data_input_stream_new(
             g_memory_input_stream_new_from_data(answer->body, remaining, NULL));
@@ -173,7 +173,7 @@ void weechat_receive(weechat_t* weechat)
     while (remaining > 0) {
         type_t type = weechat_decode_type(mem, &remaining);
         weechat_unmarshal(mem, type, &remaining);
-        g_printf("Remain: %zu\n", remaining);
+        g_printf("Remain: %d\n", remaining);
     }
 }
 
@@ -214,7 +214,7 @@ answer_t* weechat_parse_header(weechat_t* weechat)
     return answer;
 }
 
-void weechat_unmarshal(GDataInputStream* stream, type_t type, gsize* remaining)
+void weechat_unmarshal(GDataInputStream* stream, type_t type, gint32* remaining)
 {
     gchar w_chr;
     gint32 w_int;
@@ -286,7 +286,7 @@ void weechat_unmarshal(GDataInputStream* stream, type_t type, gsize* remaining)
     }
 }
 
-gchar* weechat_decode_str(GDataInputStream* stream, gsize* remaining)
+gchar* weechat_decode_str(GDataInputStream* stream, gint32* remaining)
 {
     gint32 str_len = weechat_decode_int(stream, remaining);
 
@@ -308,7 +308,7 @@ gchar* weechat_decode_str(GDataInputStream* stream, gsize* remaining)
     return g_string_free(msg, FALSE);
 }
 
-gchar weechat_decode_chr(GDataInputStream* stream, gsize* remaining)
+gchar weechat_decode_chr(GDataInputStream* stream, gint32* remaining)
 {
     gchar c = g_data_input_stream_read_byte(stream, NULL, NULL);
     *remaining -= 1;
@@ -316,7 +316,7 @@ gchar weechat_decode_chr(GDataInputStream* stream, gsize* remaining)
     return c;
 }
 
-gint32 weechat_decode_int(GDataInputStream* stream, gsize* remaining)
+gint32 weechat_decode_int(GDataInputStream* stream, gint32* remaining)
 {
     gint32 i = g_data_input_stream_read_int32(stream, NULL, NULL);
     *remaining -= 4;
@@ -324,7 +324,7 @@ gint32 weechat_decode_int(GDataInputStream* stream, gsize* remaining)
     return i;
 }
 
-gint64 weechat_decode_lon(GDataInputStream* stream, gsize* remaining)
+gint64 weechat_decode_lon(GDataInputStream* stream, gint32* remaining)
 {
     gchar length = weechat_decode_chr(stream, remaining);
 
@@ -338,7 +338,7 @@ gint64 weechat_decode_lon(GDataInputStream* stream, gsize* remaining)
     return g_ascii_strtoll(lon, NULL, 10);
 }
 
-gchar* weechat_decode_ptr(GDataInputStream* stream, gsize* remaining)
+gchar* weechat_decode_ptr(GDataInputStream* stream, gint32* remaining)
 {
     gchar length = weechat_decode_chr(stream, remaining);
 
@@ -352,7 +352,7 @@ gchar* weechat_decode_ptr(GDataInputStream* stream, gsize* remaining)
     return g_strdup_printf("0x%s", ptr);
 }
 
-gchar* weechat_decode_tim(GDataInputStream* stream, gsize* remaining)
+gchar* weechat_decode_tim(GDataInputStream* stream, gint32* remaining)
 {
     gchar length = weechat_decode_chr(stream, remaining);
 
@@ -365,7 +365,7 @@ gchar* weechat_decode_tim(GDataInputStream* stream, gsize* remaining)
     return g_string_free(msg, FALSE);
 }
 
-GVariant* weechat_decode_arr(GDataInputStream* stream, gsize* remaining)
+GVariant* weechat_decode_arr(GDataInputStream* stream, gint32* remaining)
 {
     type_t arr_t = weechat_decode_type(stream, remaining);
     gint32 arr_l = weechat_decode_int(stream, remaining);
@@ -398,7 +398,7 @@ GVariant* weechat_decode_arr(GDataInputStream* stream, gsize* remaining)
     return value;
 }
 
-GVariant* weechat_decode_inf(GDataInputStream* stream, gsize* remaining)
+GVariant* weechat_decode_inf(GDataInputStream* stream, gint32* remaining)
 {
     GVariant* pair;
 
@@ -431,7 +431,7 @@ GVariant* weechat_decode_inf(GDataInputStream* stream, gsize* remaining)
  *   ]
  * }
  */
-GVariant* weechat_decode_inl(GDataInputStream* stream, gsize* remaining)
+GVariant* weechat_decode_inl(GDataInputStream* stream, gint32* remaining)
 {
     GVariantDict* inl = g_variant_dict_new(NULL);
 
@@ -474,7 +474,7 @@ GVariant* weechat_decode_inl(GDataInputStream* stream, gsize* remaining)
 }
 
 // TODO: Test me.
-GVariant* weechat_decode_htb(GDataInputStream* stream, gsize* remaining)
+GVariant* weechat_decode_htb(GDataInputStream* stream, gint32* remaining)
 {
     GVariantBuilder* builder;
     type_t k, v;
@@ -502,7 +502,7 @@ GVariant* weechat_decode_htb(GDataInputStream* stream, gsize* remaining)
     return g_variant_builder_end(builder);
 }
 
-GVariant* weechat_decode_hda(GDataInputStream* stream, gsize* remaining)
+GVariant* weechat_decode_hda(GDataInputStream* stream, gint32* remaining)
 {
     GVariantBuilder* builder;
     gchar* path, *keys;
@@ -590,7 +590,7 @@ GVariant* weechat_decode_hda(GDataInputStream* stream, gsize* remaining)
     return g_variant_builder_end(builder);
 }
 
-type_t weechat_decode_type(GDataInputStream* stream, gsize* remaining)
+type_t weechat_decode_type(GDataInputStream* stream, gint32* remaining)
 {
     gchar type[4];
